src/transaction_fee_optimization.py
```python
from web3 import Web3
from src.config import load_config

class TransactionFeeOptimizer:
    def __init__(self):
        self.config = load_config()
        self.web3 = Web3(Web3.HTTPProvider(self.config['blockchain_network']))

    def get_gas_price(self):
        return self.web3.eth.gasPrice

    def optimize_transaction_fee(self, transaction):
        gas_price = self.get_gas_price()
        optimized_fee = self.calculate_optimized_fee(transaction, gas_price)
        return optimized_fee

    def calculate_optimized_fee(self, transaction, gas_price):
        gas_limit = self.estimate_gas_limit(transaction)
        optimized_fee = gas_price * gas_limit
        return optimized_fee

    def estimate_gas_limit(self, transaction):
        return self.web3.eth.estimateGas(transaction)

if __name__ == "__main__":
    optimizer = TransactionFeeOptimizer()
    transaction = {
        'to': '0x0',
        'value': 0,
        'gas': 0,
        'gasPrice': 0,
        'nonce': 0,
        'chainId': 0
    }
    optimized_fee = optimizer.optimize_transaction_fee(transaction)
    print(f"Optimized transaction fee: {optimized_fee}")
```