tests/test_transaction_fee_optimization.py
```python
import unittest
from unittest.mock import patch
from src.transaction_fee_optimization import optimize_transaction_fee

class TestTransactionFeeOptimization(unittest.TestCase):

    @patch('src.transaction_fee_optimization.get_market_data')
    @patch('src.transaction_fee_optimization.get_price_data')
    @patch('src.transaction_fee_optimization.execute_transaction')
    def test_optimize_transaction_fee(self, mock_execute_transaction, mock_get_price_data, mock_get_market_data):
        # Mocking the return values of the dependencies
        mock_get_market_data.return_value = {"liquidity": 1000, "price": 1, "volume": 10000}
        mock_get_price_data.return_value = 1
        mock_execute_transaction.return_value = True

        # Test case: Check if the function returns True when transaction is executed successfully
        self.assertTrue(optimize_transaction_fee('BUY', 'CAKE', 10))

        # Test case: Check if the function returns False when transaction is not executed successfully
        mock_execute_transaction.return_value = False
        self.assertFalse(optimize_transaction_fee('BUY', 'CAKE', 10))

        # Test case: Check if the function raises an exception when an invalid operation is passed
        with self.assertRaises(ValueError):
            optimize_transaction_fee('INVALID_OPERATION', 'CAKE', 10)

if __name__ == '__main__':
    unittest.main()
```